working solution - db refactoring
