+++ mpcforces_extractor/api/config.py
@@ -1,0 +1,7 @@
+ITEMS_PER_PAGE = 100
+UPLOAD_FOLDER = "data/uploads"
+OUTPUT_FOLDER = "data/output"
+STATIC_DIR = (
+    "mpcforces_extractor/visualization/frontend/static"  # Path to the static directory
+)
+TEMPLATES_DIR = "mpcforces_extractor/visualization/frontend/templates"  # Path to the templates directory
